Tubes > List fixes form submission for filter and printing tube labels
HTML form inputs like text fields and checkboxes support an HTML attribute
"form"[1] declaring which <form> tag that input should be associated with.
And that <input> does NOT need to be inside/nested within the associated
<form> tag--rather that's the convention most forms follow.

The table has a search/filter form in the <thead>, and the <tbody> uses
checkboxes in rows for the print labels form. HTML tables are very strict
with where they allow <form> tags to exist, so the best strategy was adding
<form> tags entirely outside the <table>, while the associated <input>'s are
mixed into <th> and <td> tags and given the form="" they're associated with.

[1] https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input#htmlattrdefform

CVDLS-156